Move dummyarray into the CUDA target
