+++ numba/cuda/cudadrv/devicearray.py
@@ -14,12 +14,11 @@ import numpy as np
 
 import numba
 from numba import _devicearray
-from numba.cuda.cudadrv import devices
+from numba.cuda.cudadrv import devices, dummyarray
 from numba.cuda.cudadrv import driver as _driver
 from numba.core import types, config
 from numba.np.unsafe.ndarray import to_fixed_tuple
 from numba.np.numpy_support import numpy_version
-from numba.misc import dummyarray
 from numba.np import numpy_support
 from numba.cuda.api_util import prepare_shape_strides_dtype
 from numba.core.errors import NumbaPerformanceWarning
